Allow init timeout error, move traffic controller to router pkg
