+++ aggregator/handler/router/traffic_controlled.go
@@ -21,10 +21,109 @@
 package router
 
 import (
+	"time"
+
 	"github.com/m3db/m3aggregator/aggregator/handler/common"
+	"github.com/m3db/m3cluster/kv"
+	"github.com/m3db/m3cluster/kv/util"
+	"github.com/m3db/m3x/watch"
 
 	"github.com/uber-go/tally"
+	"go.uber.org/atomic"
 )
+
+const (
+	defaultInitTimeout = 2 * time.Second
+)
+
+// TrafficController controls traffic.
+type TrafficController interface {
+	// Allow returns true if traffic is allowed.
+	Allow() bool
+
+	// Init initializes the traffic controller to watch the runtime updates.
+	Init() error
+
+	// Close closes the traffic controller.
+	Close()
+}
+
+type trafficEnabler struct {
+	enabled *atomic.Bool
+	value   watch.Value
+	opts    TrafficControlOptions
+}
+
+// NewTrafficEnabler creates a new traffic controller.
+func NewTrafficEnabler(opts TrafficControlOptions) TrafficController {
+	enabled := atomic.NewBool(opts.DefaultValue())
+	iOpts := opts.InstrumentOptions()
+	newUpdatableFn := func() (watch.Updatable, error) {
+		w, err := opts.Store().Watch(opts.RuntimeKey())
+		return w, err
+	}
+	getFn := func(updatable watch.Updatable) (interface{}, error) {
+		return updatable.(kv.ValueWatch).Get(), nil
+	}
+	processFn := func(update interface{}) error {
+		b, err := util.BoolFromValue(
+			update.(kv.Value),
+			opts.RuntimeKey(),
+			opts.DefaultValue(),
+			util.NewOptions().SetLogger(iOpts.Logger()),
+		)
+		if err != nil {
+			return err
+		}
+		enabled.Store(b)
+		return nil
+	}
+	vOptions := watch.NewOptions().
+		SetInitWatchTimeout(opts.InitTimeout()).
+		SetInstrumentOptions(iOpts).
+		SetNewUpdatableFn(newUpdatableFn).
+		SetGetUpdateFn(getFn).
+		SetProcessFn(processFn)
+	return &trafficEnabler{
+		enabled: enabled,
+		value:   watch.NewValue(vOptions),
+		opts:    opts,
+	}
+}
+
+func (c *trafficEnabler) Init() error {
+	err := c.value.Watch()
+	if err == nil {
+		return nil
+	}
+	if _, ok := err.(watch.CreateWatchError); ok {
+		return err
+	}
+	return nil
+}
+
+func (c *trafficEnabler) Close() {
+	c.value.Unwatch()
+}
+
+func (c *trafficEnabler) Allow() bool {
+	return c.enabled.Load()
+}
+
+type trafficDisabler struct {
+	TrafficController
+}
+
+// NewTrafficDisabler creates a new traffic disabler.
+func NewTrafficDisabler(opts TrafficControlOptions) TrafficController {
+	return &trafficDisabler{
+		TrafficController: NewTrafficEnabler(opts),
+	}
+}
+
+func (c *trafficDisabler) Allow() bool {
+	return !c.TrafficController.Allow()
+}
 
 type trafficControlledRouterMetrics struct {
 	trafficControlNotAllwed tally.Counter
@@ -37,7 +136,7 @@ func newTrafficControlledRouterMetrics(scope tally.Scope) trafficControlledRoute
 }
 
 type trafficControlledRouter struct {
-	common.TrafficController
+	TrafficController
 	Router
 
 	m trafficControlledRouterMetrics
@@ -45,7 +144,7 @@ type trafficControlledRouter struct {
 
 // NewTrafficControlledRouter creates a traffic controlled router.
 func NewTrafficControlledRouter(
-	trafficController common.TrafficController,
+	trafficController TrafficController,
 	router Router,
 	scope tally.Scope,
 ) Router {
